move package specific information parsing to parser_package.py
