+++ gentoo_update/__init__.py
@@ -4,6 +4,7 @@ from ._version import __version__
 from .gentoo_update import add_prefixes, create_cli, main
 from .notifier import Notifier
 from .parser import Parser
+from .parser_package import PackageParser
 from .report_objects import (
     DiskUsage,
     DiskUsageStats,
